Transfer code for exporting everything from AdminOps to ConfigOps
